e164: Standarize numbers from/to Kamtrunks to E164
The initial approach for number standarization has the following
rules:

- Kamailio Trunks <-> Application Servers
Both outgoing and incming calls we will always use E164 standarized
numbers for both origin and destination. This format will be also
used any header with a number present.

- Kamailio Users <-> Application Servers
With outgoing calls, the dialed number will be converted only if it
is an external number as follows:
 * Check Company OutboundPrefix
 * Convert dialed number to ACL format and check them
 * Convert dialed number to E164 before sending to KT

With incoming calls, the origin number will be converted from E164 to
the format preferred by the company or user (if it ends in a user
endpoint) taking into the account the company outboundPrefix, the
company country and the destination user country.