billing: Fix invoice generation that includes too many target patterns
Invoice generator expands defined templates code into a half-cooked template
that will be used to replace with the final values.

One of the templated provided as example contains tables with detail for each
of the invoice target patterns calls. If there are too many of this tables
(like 15 or more), the regexp used to create the half-cooked template produces
a segfault in pcre library.