portals: Fix spinner position in numeric input fields
This is something that was killing my insides from a log time.

Kudos to davidbreyer that gave me the hint in an issue report [1] to fix this
problem.

We solved this by loading the spinner CSS in klear initialization. This may not
be the best solution but it works by now.

[1] https://github.com/btburnett3/jquery.ui.spinner/issues/10