jobs: avoid infinite requeue of gearmand jobs
Mark the GearmanJob as completed at the start of receiver worker. This may lead
to jobs never actually finishing, but we avoid an infite loop where jobs that
crashed started over and over again.

We also reduce the limit of consecutive failures of jobs, so in case of error
we can easily detect the problem using monitoring tools.