recordings: Add support for automatic recordings rotation
We have included some configuration options in Brand and Companies edit
screens to limit disk usage in recording files. This can be used to control
the space used for each entity.

When 80% of configured limit has reached, a notification mail will be sent to
the configured address.

When 100% of configured limit has reached, oldest recordings will be removed
until the disk usage is below 100% again.

Be warned, this feature doesn't reserve space, only limits the disk usage.