as: basic queue routing support #26
This change includes the following features:

 - New route type for DDI and Extension: ->Queue
 - Added entities for Queue and Queue members in portal under company section
 - Added asterisk realtime tables for queue and queue members
 - Added dialplan contexts for calls to queues and calls from queues to members
 - Implemented basic queue features
    * Queue size limit and Queue full handlers
    * Queue max time and No answer handlers
    * Basic queue strategies
    * Members priorities

This changes are still experimental. There are very binded to interface states
that right now are distributed between application servers, so they aren't as
reliable as they should for queues. When using queues, It's highly adivisable to
assign an application server to the company while prefered Application Server
features are being developed.