agi: error notifications over progress feature
This introduces a new Brand and Company feature 'Voice Notifications'. When
enabled, some errors will play a locution to the caller over progress (not
answering the call).

This provides extra information to the administrators in order to detect why
a call is not being processed.

It may also introduce some other weird scenarios where a user fordwards its
call to a destination that can not actually call and the locution is played to
the caller rather than the user who configured the call fordward (at least
provides extra information of why the fordward is not being done).