as: use brand notification address for received faxes
When a Fax is received and the send email option is enabled, the PBX will send
a notification with the fax converted to PDF attached.

The goal of this change is to use the email notification name and address
configurable at brand screen to customize email From header. We have also used
the template logics like we already do for voicemail notifications.

This also removes the extra mail notification that was being sent when an
outgoing fax failed to be sent as it's already displayed in the screen that is
used to send the fax. If this is not clear enough we can change the description
of the send email Fax field to 'send received fax to email'.