fax: replace send fax AMI request for ARI request
Now that the PDF files of the faxes are converted to TIF at PBX levels, the
request to send faxes from the portal are a merely request to one of the
running asterisks.

Asterisk provides a REST API that can be handy to remove the worker that was
handling this request: now portal can request itself to the asterisk.

This commit remove all unnecesary Gearmand and AMI connectors and replace them
by a new ARI connector to send Asterisk any request.