agi: use caller DDI for not diverted outgoing calls
When a call is being routed out of the system and no diversion header is present
the origin of the call is kept. This may clash with some carrier logics that
forces presentation from a known DDI.

This is far from the perfect. If an external call routes to some objects that
has as handler and external number, the final caller to that number will be
removed and the original DDI dialed will be the new caller. Destination will
received a call from the original called DDI, instead of a call deflected from
the original DDI.