trunks+as: bounced faxes are not negociating T.38.
Proxytrunks implements a logic that avoids forwarding medialiberation re-invites
in bounced calls. The reason behind this logic was that forwarding this re-invites
lead to a loop (as they were forwarded to AS, that was the source of this requests).

This commits deletes this logic and avoids this loop with a different approach:

- 'proxybouncer' pjsip endpoint is used to bounce calls.

- This new endpoint has direct_media to 'no', stopping the loop.

Deleted logic prevented T.38 reinvites too, so this change fixes #99 too.