agi: check diversion headers on external calls
This commit includes a new logic for all externals calls.
When a diversion header is detected, the number part of the URI will be checked.
If it matches a company extension, it's user outgoing ddi will replace the number.
If it matches a company ddi, will be kept untouched.
If it doesn't match any of these, the header will be removed.