agi: implemented IVRCustom success locution and entry welcome
This locutions where implmented in the portal interface but was not being
actually used. This changes also includes an improvement in playback wrapper
function to ignore playback requests on not configured locutions.

IVRCustom Success locution can be used to indicate a generic message to confirm
the entered input has been accepted. IVRCustom Entry Welcome locution can be
used to provide specific information what destination is being contacted.

Take into account that if both locutions are specified, both will be played.