voicemail: use context when checking destination email
Voicemail handler for sending notifications was only checking the voicemail
number but not the context, sending the nofications to first user found with
same voicemail number that the expected destination.

This changes includes the context in voicemail notification so we can get the
destination user properly.

There's also included a couple of minor fixes, formatting cleanup and extra
checks.