users: Fix NOTIFY entity field
'entity' fields of sent NOTIFY requests were incorrect as $terminal@$domain
was used instead of $extension@$domain as requested.

This was because of a dbaliases operation that was made when handling corresponding
SUBSCRIBE requests.

This situation (non-matching entity in XML body) causes problems in some hardphones
(e.g. Cisco). Other phones did not care and worked correctly.

This commit fixes in stable version what #143 does for next release (without a delta
requirement)