users+trunks: Inactivate even last AS
Skip discarding last AS was a mistake, even in standalone installations. This
decision masks a problem and prevents potential self-healing logics
(e.g. trigger some event when a AS is discarded so that it gets back asap).

With current configuration, when dispatcher module marks a GW as inactive,
it starts sending OPTIONS every 5 seconds. Retransmissions of those OPTIONS
mean that inactive GWs are monitored at least once every 1,5 seconds.

This commit reverts this decision made in #64.