recordings: fix duration and date for created recordings
This commit changes the way recordings duration is
calculated. Insted of using the created and modified
dates of the file, we let zend-mp3[0] library to
do the task.

We have also changed the recording date to match
the call start-time instead of using the current
timestamp. Note that this will set all on-demand
recordings within the same call with the same
timestamp.

[0] https://github.com/debach/zend-mp3.git