agis: restore origin E.164 number on 302 forwards
When an external call is received from proxytrunk its format is always
has E.164 format, but if the call is routed to a user, it's transformaed
by application server in what we call preferred format.

When the user's terminal has call forward enabled and a 302 moved is received,
the origin number is in user preferred format. If the redirection must be routed
externally (through proxytrunks) we must restore the E.164 format (that was
originally sent by proxytrunks when the call entered in first instance).

This change is no required in a future where application server only handles
E.164 numbers.