license: switch from EUPLv1.1 to GPLv3 License
EUPL v1.2 was released on May 2017. Our initial plan was to switch
as soon as it was published with the hope that some of the new license
compatibilities will be dropped to ensure the strong copyleft.

Sadly, this is not the case, and EUPL still has some compatible licenses
that can't ensure the forever freedom of the licensed software.

Most of the software used and installed with IvozProvider is under
General Public License version 2 or version 3, so the most coherent
free license to use for the whole software could be GPLv3.

This change can be attained directly if all project authors agree, or
via a two step licensing through CeCILL v2.