agi: replace call origin identification
Before this change, the caller of a channel was being passed from action
to action in a protected class attribute.

With this commit, both origin and caller are stored as channel variables
and retrieved from whatever action requires special processing for them.
This allow properly handling origin and callers outgoing DDI rules based on
destination.

This also includes a change in how Extensions -> Number works. Before, this
route was always handled as a DDI placing external calls. Now, based on the
caller (now stored in the channel), it routes to one of the available external
call actions (User,DDI,Friend,Retail,...) allowing custom checks and
processes.