trunks: Enable dialog cseq tracking mechanism
Some time ago we disabled dialog track_cseq_updates modparam.

This setting aims to adapt cseq numbers one Kamailio's UAC uac_auth is done.

This mechanism had some errors with retransmissions and cancel events, so it was
disabled as no carrier at the moment seems to be annoyed if two transactions use
the same CSeq.

Even though any carrier at the moment seems to be annoyed if two transactions
use the same CSeq for the unauth initial INVITE and the auth one, adapting these
CSeqs is more RFC-compliant.

We enable the mechanism again.