ivr: fix IVRCustom handlers for timeout and error
Fixed a bug where the IVRCustom handlers were not properly processed.

The logic stored the IVR id but not the IVR type. While processing the handlers
for a IVRCustom, if a IVRCommon with the same id was found they were processed
instead of the original IVRCustom handlers.

This commit includes an extra channel variable IVRTYPE to determine if the
IVR handlers being processed belong to a IVRCustom or a IVRCommon.