as: Disable direct media for static endpoints
- proxytrunks: lots of carriers show problems processing the liberation reINVITE
as soon as the session is established. Although it is 100% RFC-complaint, sadly
it is not so world-complaint.

- replacer: both endpoints send the liberation reinvite as the session is established,
causing 491 Request pending. Although this mechanism works, it is simpler not to
liberate media in this few calls.

With this commit, AS-scaling gets much more important than it was, as media of
external calls will always be handled in Asterisk. Standalone installations have
never been production-ready, but this commit makes them even less suitable.

Media handling of internal calls will continue being delegated in media-relays,
making possible to have a near media-relay to avoid latencies in internal calls.

Note that all call media will use media-relays, the only difference is that Asterisk
will stay in media-path for external calls.