agi: make origin/caller channel variables inheritable
When we receive a terminal call forward, asterisk will create
a Local/ channel with redirection information. We can get the caller
from Diversion data, but the origin is lost from this process.

This change makes CALLER and ORIGIN variables to be inherited in
diverted channel, so we can set properly outgoingDDI if required