kamtrunks: Bounce to same AS that asks for it
Bouncing mechanism must vanish. But in the meantime, make KamTrunks bounce
the call to the same AS that has asked that bouncing.

This way, if you free one AS assigning companies to other Asterisks, bouncing
calls are not sent to that AS (it was happening as avp(asAddress) was not set).