kam: New default distribution method (hash based)
Introduce a new algorithm to distribute calls between Application Servers: hash
based. With this algorithm, all calls of a company/retail are handled in the same
AS. If this AS gets unreachable, another one is used.

This algorithm is the recommended (and default) one for all companies, and
round-robin is reserved for huge companies whose calls cannot be handled in a
single AS. Round-robin is not compatible with Queues.

This fixes #226