webadmin: fix available terminals listbox
When creating a new user or editing an existing one
the listbox must only display available terminals and the
one already assigned to the user.

This commit removes the function from repository and
implements it in the filter as its code is not expected
to be reused in the near future.