trunks: Reorder logics to avoid dlg_var lost on 422
Commit 938d911 introduced a new behaviour: non-2XX response forward to AS, instead
of managing it in proxytrunks.

The reason behind this was that ST is an end-to-end mechanism and it was easier
to forward it to AS and make it re-start from the beginning.

As a side effect, dlg_manage() is called twice for this calls: this is the first
time that happens and has bad consequences (dlg_vars are lost causing bad cdr
inserts that lead to unbillable calls).

This fix avoids this problem, reordering route logics to avoid this dlg_var lost.