scheme: merge IVRCommon and IVRCustom into IVRs
This migration will merge IVRCommon and IVRCustom tables into
IVRs table, updating all foreign keys.

This will also merge 'IVRCommon' and 'IVRCustom' routable
options in the enum to 'ivr', updating the tables that could
route to IVRs.

A new table has been created to store extensions exceptions
that won't be allowed to be called from the IVR even when the
allowExtensions flags is enabled.