scheme+kamusers: New table for KamUsers stats
kam_acc_cdrs had cdrs from proxies. This commit changes this.

KamUsers now saves its cdrs into a separate table is identical to kam_acc_cdrs
without unnecesary fields. It only add one/four field(s):

- owner: saved which user (friend or retailAccount) is behind each stats. It is
an initial letter (u for users, f for friends and r for retailAccounts) and its id.

There is a MySQL trigger that gets that id and sets in corresponding field (userId,
friendId, retailClientId).

NOTE: MySQL trigger is necessary as Kamailio acc only saves dlg_vars (that only
store strings). This means that it cannot save only userId and leave remaining values
(friendId, retailAccountId) to NULL. It would save '' that makes MySQL not to save
the row as it is not a valid value for an int field.