core: update Endpoint/Voicemail when Extension number changes
Ast/PsEndpoint and Ast/Voicemail were only being updated when
the User changed the extension. This two tables contain the extension
number, so must be updated if the Extension itself changes, not only
its relation with the user.

This change only applies to those extension that are used as screen
extension for the user, not all the extension that route to user.