kam+agis: Fix misleading maxcalls implementation
This commit fixes the misleading implementation made in #398 for maxcalls.

Before this commit:

- KamUsers had an independent counter that counted dialogs in both directions:
  - UAC -> KamUsers -> AS
  - AS -> KamUsers -> UAC

- KamTrunks had an independent counter that counted dialogs in both directions:
  - Carrier -> KamTrunks -> AS
  - AS -> KamTrunks -> Carrier

This made difficult to answer the question "how many calls can I handle with this
setting set to 10":

- 10 external outgoing calls made by users.

or

- 10 external incoming calls received by users.

or

- 10 external incoming calls forwarded outside again + 10 internal calls.

After this commit:

- KamUsers has an independent counter but only counts (and limits) UAC->AS calls.
- KamTrunks has an independent counter but only counts (and limits) Carrier->AS calls.

This way the previous question can be answered easily:

- 10 calls generated by users + 10 calls received from the outside world.

Until a counter is shared between Kamusers and KamTrunks, this implementation seems
less ambiguous.