doc: reworked documentation tree and add links in every list screen
This commit changes the documentation order to look more like the
administration portal menu, making it easier to find each section
documentation.

Most of the work is just moving files and translations to the new
groups and it still lot of rework in order to fill all new sections
and remove/replace deprecated information from older releases.