kam+as: Redesign bouncing logic
This commit aims to remove all the complex bounce logic. For that, it disables
detect_spiral dialog modparam so that a bounced call gets all the dialog info
structure recreated.

This way, a bounced call will make 2 CDR inserts, one as an outbound call and
the other as an inbound call.

When KamTrunks detects a call is for a DDI hosted in its own database, it sends
the request to himself, triggering the bouncing logic.

This commit also introduces X-Info-Type header to classify calls between vpbx,
retail and wholesale, instead of sending X-Info-WholesaleId.