kamtrunks: warn CGRateS of call failure in failover
Instead of calling CGR_CALL_FAILED in dialog:failed, that is only executed when
the last tried carrier fails, call CGR_CALL_FAILED in failure_route.

Get another cgrid for failing call mangling from_tag, so that the original cgrid
is reserved for the eventual success of the call (using other carriers).

This commit fixes #647.