kamusers: manual detect spirals
Retail calls to DDIs of another (or same) retail client make a bouncing
call without any AS in the signalling path. As in kamtrunks bouncing mechanism,
two CDRs are needed: one for the caller client, one for the called client.

For this, detect_spirals must be disabled and dlg_manage() must be called twice.

To limit this double call to this scenario (and avoid dlg_var loss), is_known_dlg()
and request source are verified.