agis: fix call forward settings for routes to voicemail
Code was assuming the voicemail entity was always a UserInterface.
This is not the case for Call Forwards, that are used by both
Users and ResidentialDevices.

The following commit updates all routes to treat the voicemail as
a UserInterface, except Call Forward settings that will choose
between UserInterface or ResidentialDevice.