kamtrunks: do not inactivate stopper gateways
A carrier may be used both in stopper rule and in non-stopper one.

With previous logic, non stopper rule may inactivate a carrier that is stopper
in another rule. As inactivated gateways are not taken into account in new calls,
stopper rule may be skipped.

This commit avoids inactivating a carrier that is stopper in any rule.