kamusers: increase presence clean_period
clean_period was set to 15 instead of default 100 value in f6addae35 (PR #1570)
to have BLF light off asap in case of rejected call establishments.

To avoid extensive database usage, this value goes back to its default value of 100.
With this change, BLF light will be off a bit later (not 3 hours as the Kamailio
patch persists, but neither 30 seconds... new intermediate reasonable value: 2-3 minutes)

It is a valid value because this reject call establishment are quite rare and database
access is expensive.