kamtrunks: fix carrier/ddiprovider mediarelay selection
route(RTPENGINE) route needs dlg_var(mediaRelaySetId) to be set in order to assign proper media-relay to each call depending on which carrier/ddiprovider is involved.

With previous logic, it was used before being assigned, so it failed.

Besides, on _failure_route_ was not invoked again (to choose proper media-relay of eventual failover carrier).

This commit moves RTPENGINE call to branch_route, so that mediarelayset is properly set in every situation.