microservices/realtime: send call events in a single goroutine
Before this change, when an new websocket connection was received
a goroutine was being generated to handle it. Inside this goroutine,
a new goroutine was generated to handle challenge and subscribe key,
and another goroutine to send received events in an infinite loop.

This change simplifies two sections of the logic:
 1. Only a single goroutine for handling event retransmission.
     Using so many goroutines caused syncronization problems while
     writing data to the socket. Also, there is no need to goroutine
     websocket server handler as it is already designed for async
     handling multiple connections.

 2. Send all events from redis to websocket client:
     This is a change from the original source code but seems to
     work ok. All events generated from SIP proxies are now sent
     to web portal, not just the "significant call events".