schema: fix ps endpoint hints migration
This migration updated ast_ps_endpoints table with the appropiate
hint_extension and subscribe_context values, but it was only
considering terminals with user and extension assigned.

This is ok for hint_extension but subscribe_context must be updated
in all endpoints, no matter if there is user or not using that
terminal.