microservices/recordings: use hashid for temporal MP3 filename
Before this change, Call-Id header value was used for the
temporal mp3 file of recordings. This value can contain characters
that breaks the avconv command (like '/') so, we could encode this
value or just replace it for another one. This is a temporal file
that will be deleted by FSO core logic, so the easiest was to change
its value to hashid, a partial md5 of Call-Id.